+++ src/onegov.gazette/onegov/gazette/tests/test_views_dashboard.py
@@ -22,6 +22,7 @@ def test_view_dashboard(gazette_app):
         # new notice
         manage = manage.click("Neu")
         manage.form['title'] = "Erneuerungswahlen"
+        manage.form['organization'] = '100'
         manage.form['category'] = '1403'
         manage.form['issues'] = ['2017-44', '2017-45']
         manage.form['text'] = "1. Oktober 2017"