+++ src/onegov.gazette/onegov/gazette/tests/test_views_notices.py
@@ -23,6 +23,7 @@ def test_view_notices(gazette_app):
         # new notice
         manage = client.get('/notices/drafted/new-notice')
         manage.form['title'] = "Erneuerungswahlen"
+        manage.form['organization'] = '210'
         manage.form['category'] = '1403'
         manage.form['issues'] = ['2017-44', '2017-45']
         manage.form['text'] = "1. Oktober 2017"
@@ -77,6 +78,7 @@ def test_view_notices_search(gazette_app):
         # new notice
         manage = client.get('/notices/drafted/new-notice')
         manage.form['title'] = "Erneuerungswahlen"
+        manage.form['organization'] = '210'
         manage.form['category'] = '1403'
         manage.form['issues'] = ['2017-44', '2017-45']
         manage.form['text'] = "1. Oktober 2017"
@@ -86,6 +88,7 @@ def test_view_notices_search(gazette_app):
 
         manage = client.get('/notices/drafted/new-notice')
         manage.form['title'] = "Kantonsratswahlen"
+        manage.form['organization'] = '210'
         manage.form['category'] = '1403'
         manage.form['issues'] = ['2017-44', '2017-45']
         manage.form['text'] = "10. Oktober 2017"
@@ -116,6 +119,10 @@ def test_view_notices_statistics(gazette_app):
     publisher = Client(gazette_app)
     login_publisher(publisher)
 
+    url_organizations = '/notices/{}/statistics-organizations'
+    url_categories = '/notices/{}/statistics-categories'
+    url_groups = '/notices/{}/statistics-groups'
+
     # No notices yet
     states = ('drafted', 'submitted', 'published', 'rejected')
     for s in states:
@@ -124,37 +131,45 @@ def test_view_notices_statistics(gazette_app):
         editor.get('/notices/{}/statistics-categories'.format(s), status=403)
 
         publisher.get('/notices/{}/statistics'.format(s))
-        assert publisher.get(
-            '/notices/{}/statistics-groups'.format(s)
-        ).text == 'Gruppe,Anzahl\r\n'
-        publisher.get(
-            '/notices/{}/statistics-categories'.format(s)
-        ).text == (
-            'Rubrik,Titel,Anzahl\r\n'
-            '12,Weiterbildung,0\r\n'
-            '13,Submissionen,0\r\n'
-            '14,Kantonale Mitteilungen,0\r\n'
-            '1402,Kantonale Mitteilungen / Einberufung Kantonsrat,0\r\n'
-            '1403,Kantonale Mitteilungen / Wahlen/Abstimmungen,0\r\n'
-            '1406,Kantonale Mitteilungen / Kant. Gesetzgebung,0\r\n'
-            '1411,Kantonale Mitteilungen / Mitteilungen Landschreiber,0\r\n'
-            '1412,Kantonale Mitteilungen / Kant. Stellenangebote,0\r\n'
-            '1413,Kantonale Mitteilungen / Direktion des Innern,0\r\n'
-            '1414,Kantonale Mitteilungen / Direktion für Bildung und Kultur,0'
-            '\r\n'
-            '1415,Kantonale Mitteilungen / Volkswirtschaftsdirektion,0\r\n'
-            '1416,Kantonale Mitteilungen / Baudirektion,0\r\n'
-            '1418,Kantonale Mitteilungen / Gesundheitsdirektion,0\r\n'
-            '1421,Kantonale Mitteilungen / Finanzdirektion,0\r\n'
-            '1426,Kantonale Mitteilungen / Gerichtliche Bekanntmachungen,0\r\n'
-            '1427,Kantonale Mitteilungen / Konkursamt,0\r\n'
-            '16,Bürgergemeinden,0\r\n'
-            '17,Kath. Kirchgemeinden,0\r\n'
-            '18,Ev.-ref. Kirchgemeinde,0\r\n'
-            '19,Korporationen,0\r\n'
-            '20,Handelsregister,0\r\n'
+        assert publisher.get(url_organizations.format(s)).text == (
+            'Organisation,Anzahl\r\n'
+            'Staatskanzlei Kanton Zug,0\r\n'
+            'Bürgergemeinde Zug,0\r\n'
+            'Einwohnergemeinde Zug,0\r\n'
+            'Evangelisch-reformierte Kirchgemeinde des Kantons Zug,0\r\n'
+            'Katholische Kirchgemeinde Baar,0\r\n'
+            'Katholische Kirchgemeinde Zug,0\r\n'
+            'Korporation Zug,0\r\n'
+        )
+        assert publisher.get(url_categories.format(s)).text == (
+            'Rubrik,Anzahl\r\n'
+            'Bürgergemeinden,0\r\n'
+            'Ev.-ref. Kirchgemeinde,0\r\n'
+            'Handelsregister,0\r\n'
+            'Kantonale Mitteilungen,0\r\n'
+            'Kantonale Mitteilungen / Baudirektion,0\r\n'
+            'Kantonale Mitteilungen / Direktion des Innern,0\r\n'
+            'Kantonale Mitteilungen / Direktion für Bildung und Kultur,0\r\n'
+            'Kantonale Mitteilungen / Einberufung Kantonsrat,0\r\n'
+            'Kantonale Mitteilungen / Finanzdirektion,0\r\n'
+            'Kantonale Mitteilungen / Gerichtliche Bekanntmachungen,0\r\n'
+            'Kantonale Mitteilungen / Gesundheitsdirektion,0\r\n'
+            'Kantonale Mitteilungen / Kant. Gesetzgebung,0\r\n'
+            'Kantonale Mitteilungen / Kant. Stellenangebote,0\r\n'
+            'Kantonale Mitteilungen / Konkursamt,0\r\n'
+            'Kantonale Mitteilungen / Mitteilungen Landschreiber,0\r\n'
+            'Kantonale Mitteilungen / Volkswirtschaftsdirektion,0\r\n'
+            'Kantonale Mitteilungen / Wahlen/Abstimmungen,0\r\n'
+            'Kath. Kirchgemeinden,0\r\n'
+            'Korporationen,0\r\n'
+            'Submissionen,0\r\n'
+            'Weiterbildung,0\r\n'
+        )
+        assert publisher.get(url_groups.format(s)).text == (
+            'Gruppe,Anzahl\r\n'
         )
 
+    # Add users and groups
     admin = Client(gazette_app)
     login_admin(admin)
     manage = admin.get('/groups').click("Neu")
@@ -193,24 +208,25 @@ def test_view_notices_statistics(gazette_app):
 
     # Add notices
     with freeze_time("2017-11-01 12:00"):
-        for (category, submit, user) in (
-            ('13', False, editor),
-            ('13', False, user_1),
-            ('1406', False, user_1),
-            ('1406', False, user_1),
-            ('16', True, user_1),
-            ('19', True, user_1),
-            ('19', True, user_1),
-            ('14', False, user_2),
-            ('16', True, user_2),
-            ('19', False, user_2),
-            ('19', True, user_3),
-            ('16', True, user_3),
-            ('19', False, user_3),
-            ('19', True, user_3),
+        for (organization, category, submit, user) in (
+            ('100', '13', False, editor),
+            ('100', '13', False, user_1),
+            ('100', '1406', False, user_1),
+            ('210', '1406', False, user_1),
+            ('100', '16', True, user_1),
+            ('100', '19', True, user_1),
+            ('310', '19', True, user_1),
+            ('100', '14', False, user_2),
+            ('100', '16', True, user_2),
+            ('210', '19', False, user_2),
+            ('100', '19', True, user_3),
+            ('100', '16', True, user_3),
+            ('100', '19', False, user_3),
+            ('100', '19', True, user_3),
         ):
             manage = user.get('/notices/drafted/new-notice')
             manage.form['title'] = "Titel"
+            manage.form['organization'] = organization
             manage.form['category'] = category
             manage.form['issues'] = ['2017-44']
             manage.form['text'] = "Text"
@@ -219,43 +235,131 @@ def test_view_notices_statistics(gazette_app):
                 manage.click("Einreichen").form.submit()
 
     for s in ('rejected', 'published'):
+        assert publisher.get(url_organizations.format(s)).text == (
+            'Organisation,Anzahl\r\n'
+            'Staatskanzlei Kanton Zug,0\r\n'
+            'Bürgergemeinde Zug,0\r\n'
+            'Einwohnergemeinde Zug,0\r\n'
+            'Evangelisch-reformierte Kirchgemeinde des Kantons Zug,0\r\n'
+            'Katholische Kirchgemeinde Baar,0\r\n'
+            'Katholische Kirchgemeinde Zug,0\r\n'
+            'Korporation Zug,0\r\n'
+        )
         assert "Bürgergemeinden" in publisher.get(
             '/notices/{}/statistics'.format(s)
         )
-        assert publisher.get(
-            '/notices/{}/statistics-groups'.format(s)
-        ).text == 'Gruppe,Anzahl\r\nA,0\r\nB,0\r\nC,0\r\n'
-        publisher.get(
-            '/notices/{}/statistics-categories'.format(s)
-        ).text == (
-            'Rubrik,Titel,Anzahl\r\n'
-            '12,Weiterbildung,0\r\n'
-            '13,Submissionen,0\r\n'
-            '14,Kantonale Mitteilungen,0\r\n'
-            '1402,Kantonale Mitteilungen / Einberufung Kantonsrat,0\r\n'
-            '1403,Kantonale Mitteilungen / Wahlen/Abstimmungen,0\r\n'
-            '1406,Kantonale Mitteilungen / Kant. Gesetzgebung,0\r\n'
-            '1411,Kantonale Mitteilungen / Mitteilungen Landschreiber,0\r\n'
-            '1412,Kantonale Mitteilungen / Kant. Stellenangebote,0\r\n'
-            '1413,Kantonale Mitteilungen / Direktion des Innern,0\r\n'
-            '1414,Kantonale Mitteilungen / Direktion für Bildung und Kultur,0'
-            '\r\n'
-            '1415,Kantonale Mitteilungen / Volkswirtschaftsdirektion,0\r\n'
-            '1416,Kantonale Mitteilungen / Baudirektion,0\r\n'
-            '1418,Kantonale Mitteilungen / Gesundheitsdirektion,0\r\n'
-            '1421,Kantonale Mitteilungen / Finanzdirektion,0\r\n'
-            '1426,Kantonale Mitteilungen / Gerichtliche Bekanntmachungen,0\r\n'
-            '1427,Kantonale Mitteilungen / Konkursamt,0\r\n'
-            '16,Bürgergemeinden,0\r\n'
-            '17,Kath. Kirchgemeinden,0\r\n'
-            '18,Ev.-ref. Kirchgemeinde,0\r\n'
-            '19,Korporationen,0\r\n'
-            '20,Handelsregister,0\r\n'
+        assert publisher.get(url_categories.format(s)).text == (
+            'Rubrik,Anzahl\r\n'
+            'Bürgergemeinden,0\r\n'
+            'Ev.-ref. Kirchgemeinde,0\r\n'
+            'Handelsregister,0\r\n'
+            'Kantonale Mitteilungen,0\r\n'
+            'Kantonale Mitteilungen / Baudirektion,0\r\n'
+            'Kantonale Mitteilungen / Direktion des Innern,0\r\n'
+            'Kantonale Mitteilungen / Direktion für Bildung und Kultur,0\r\n'
+            'Kantonale Mitteilungen / Einberufung Kantonsrat,0\r\n'
+            'Kantonale Mitteilungen / Finanzdirektion,0\r\n'
+            'Kantonale Mitteilungen / Gerichtliche Bekanntmachungen,0\r\n'
+            'Kantonale Mitteilungen / Gesundheitsdirektion,0\r\n'
+            'Kantonale Mitteilungen / Kant. Gesetzgebung,0\r\n'
+            'Kantonale Mitteilungen / Kant. Stellenangebote,0\r\n'
+            'Kantonale Mitteilungen / Konkursamt,0\r\n'
+            'Kantonale Mitteilungen / Mitteilungen Landschreiber,0\r\n'
+            'Kantonale Mitteilungen / Volkswirtschaftsdirektion,0\r\n'
+            'Kantonale Mitteilungen / Wahlen/Abstimmungen,0\r\n'
+            'Kath. Kirchgemeinden,0\r\n'
+            'Korporationen,0\r\n'
+            'Submissionen,0\r\n'
+            'Weiterbildung,0\r\n'
+        )
+        assert publisher.get(url_groups.format(s)).text == (
+            'Gruppe,Anzahl\r\n'
+            'A,0\r\n'
+            'B,0\r\n'
+            'C,0\r\n'
         )
 
-    # drafted: 1 x w/o, 5 x B, 1 x C
+    # organizations/drafted: 5 x 100, 2 x 210
+    assert publisher.get(url_organizations.format('drafted')).text == (
+        'Organisation,Anzahl\r\n'
+        'Staatskanzlei Kanton Zug,5\r\n'
+        'Bürgergemeinde Zug,2\r\n'
+        'Einwohnergemeinde Zug,0\r\n'
+        'Evangelisch-reformierte Kirchgemeinde des Kantons Zug,0\r\n'
+        'Katholische Kirchgemeinde Baar,0\r\n'
+        'Katholische Kirchgemeinde Zug,0\r\n'
+        'Korporation Zug,0\r\n'
+    )
+
+    # organizations/submitted: 6 x 100, 1 x 310
+    assert publisher.get(url_organizations.format('submitted')).text == (
+        'Organisation,Anzahl\r\n'
+        'Staatskanzlei Kanton Zug,6\r\n'
+        'Bürgergemeinde Zug,0\r\n'
+        'Einwohnergemeinde Zug,1\r\n'
+        'Evangelisch-reformierte Kirchgemeinde des Kantons Zug,0\r\n'
+        'Katholische Kirchgemeinde Baar,0\r\n'
+        'Katholische Kirchgemeinde Zug,0\r\n'
+        'Korporation Zug,0\r\n'
+    )
+
+    # categories/drafted: 2 x 13, 2 x 1406, 1 x 14, 2 x 19
+    assert '>2</td>' in publisher.get('/notices/drafted/statistics')
+    assert publisher.get(url_categories.format('drafted')).text == (
+        'Rubrik,Anzahl\r\n'
+        'Bürgergemeinden,0\r\n'
+        'Ev.-ref. Kirchgemeinde,0\r\n'
+        'Handelsregister,0\r\n'
+        'Kantonale Mitteilungen,1\r\n'
+        'Kantonale Mitteilungen / Baudirektion,0\r\n'
+        'Kantonale Mitteilungen / Direktion des Innern,0\r\n'
+        'Kantonale Mitteilungen / Direktion für Bildung und Kultur,0\r\n'
+        'Kantonale Mitteilungen / Einberufung Kantonsrat,0\r\n'
+        'Kantonale Mitteilungen / Finanzdirektion,0\r\n'
+        'Kantonale Mitteilungen / Gerichtliche Bekanntmachungen,0\r\n'
+        'Kantonale Mitteilungen / Gesundheitsdirektion,0\r\n'
+        'Kantonale Mitteilungen / Kant. Gesetzgebung,2\r\n'
+        'Kantonale Mitteilungen / Kant. Stellenangebote,0\r\n'
+        'Kantonale Mitteilungen / Konkursamt,0\r\n'
+        'Kantonale Mitteilungen / Mitteilungen Landschreiber,0\r\n'
+        'Kantonale Mitteilungen / Volkswirtschaftsdirektion,0\r\n'
+        'Kantonale Mitteilungen / Wahlen/Abstimmungen,0\r\n'
+        'Kath. Kirchgemeinden,0\r\n'
+        'Korporationen,2\r\n'
+        'Submissionen,2\r\n'
+        'Weiterbildung,0\r\n'
+    )
+
+    # categories/submitted: 3x16, 4x19
+    assert '>3</td>' in publisher.get('/notices/submitted/statistics')
+    assert publisher.get(url_categories.format('submitted')).text == (
+        'Rubrik,Anzahl\r\n'
+        'Bürgergemeinden,3\r\n'
+        'Ev.-ref. Kirchgemeinde,0\r\n'
+        'Handelsregister,0\r\n'
+        'Kantonale Mitteilungen,0\r\n'
+        'Kantonale Mitteilungen / Baudirektion,0\r\n'
+        'Kantonale Mitteilungen / Direktion des Innern,0\r\n'
+        'Kantonale Mitteilungen / Direktion für Bildung und Kultur,0\r\n'
+        'Kantonale Mitteilungen / Einberufung Kantonsrat,0\r\n'
+        'Kantonale Mitteilungen / Finanzdirektion,0\r\n'
+        'Kantonale Mitteilungen / Gerichtliche Bekanntmachungen,0\r\n'
+        'Kantonale Mitteilungen / Gesundheitsdirektion,0\r\n'
+        'Kantonale Mitteilungen / Kant. Gesetzgebung,0\r\n'
+        'Kantonale Mitteilungen / Kant. Stellenangebote,0\r\n'
+        'Kantonale Mitteilungen / Konkursamt,0\r\n'
+        'Kantonale Mitteilungen / Mitteilungen Landschreiber,0\r\n'
+        'Kantonale Mitteilungen / Volkswirtschaftsdirektion,0\r\n'
+        'Kantonale Mitteilungen / Wahlen/Abstimmungen,0\r\n'
+        'Kath. Kirchgemeinden,0\r\n'
+        'Korporationen,4\r\n'
+        'Submissionen,0\r\n'
+        'Weiterbildung,0\r\n'
+    )
+
+    # groups/drafted: 1 x w/o, 5 x B, 1 x C
     assert '>5</td>' in publisher.get('/notices/drafted/statistics')
-    assert publisher.get('/notices/drafted/statistics-groups').text == (
+    assert publisher.get(url_groups.format('drafted')).text == (
         'Gruppe,Anzahl\r\n'
         ',1\r\n'
         'A,0\r\n'
@@ -263,65 +367,11 @@ def test_view_notices_statistics(gazette_app):
         'C,1\r\n'
     )
 
-    # submitted: 4 x B, 3 x C
+    # groups/submitted: 4 x B, 3 x C
     assert '>4</td>' in publisher.get('/notices/submitted/statistics')
-    assert publisher.get('/notices/submitted/statistics-groups').text == (
+    assert publisher.get(url_groups.format('submitted')).text == (
         'Gruppe,Anzahl\r\n'
         'A,0\r\n'
         'B,4\r\n'
         'C,3\r\n'
-    )
-
-    # drafted: 2 x 13, 2 x 1406, 1 x 14, 2 x 19
-    assert '>2</td>' in publisher.get('/notices/drafted/statistics')
-    assert publisher.get('/notices/drafted/statistics-categories').text == (
-        'Rubrik,Titel,Anzahl\r\n'
-        '12,Weiterbildung,0\r\n'
-        '13,Submissionen,2\r\n'
-        '14,Kantonale Mitteilungen,1\r\n'
-        '1402,Kantonale Mitteilungen / Einberufung Kantonsrat,0\r\n'
-        '1403,Kantonale Mitteilungen / Wahlen/Abstimmungen,0\r\n'
-        '1406,Kantonale Mitteilungen / Kant. Gesetzgebung,2\r\n'
-        '1411,Kantonale Mitteilungen / Mitteilungen Landschreiber,0\r\n'
-        '1412,Kantonale Mitteilungen / Kant. Stellenangebote,0\r\n'
-        '1413,Kantonale Mitteilungen / Direktion des Innern,0\r\n'
-        '1414,Kantonale Mitteilungen / Direktion für Bildung und Kultur,0\r\n'
-        '1415,Kantonale Mitteilungen / Volkswirtschaftsdirektion,0\r\n'
-        '1416,Kantonale Mitteilungen / Baudirektion,0\r\n'
-        '1418,Kantonale Mitteilungen / Gesundheitsdirektion,0\r\n'
-        '1421,Kantonale Mitteilungen / Finanzdirektion,0\r\n'
-        '1426,Kantonale Mitteilungen / Gerichtliche Bekanntmachungen,0\r\n'
-        '1427,Kantonale Mitteilungen / Konkursamt,0\r\n'
-        '16,Bürgergemeinden,0\r\n'
-        '17,Kath. Kirchgemeinden,0\r\n'
-        '18,Ev.-ref. Kirchgemeinde,0\r\n'
-        '19,Korporationen,2\r\n'
-        '20,Handelsregister,0\r\n'
-    )
-
-    # submitted: 3x16, 4x19
-    assert '>3</td>' in publisher.get('/notices/submitted/statistics')
-    assert publisher.get('/notices/submitted/statistics-categories').text == (
-        'Rubrik,Titel,Anzahl\r\n'
-        '12,Weiterbildung,0\r\n'
-        '13,Submissionen,0\r\n'
-        '14,Kantonale Mitteilungen,0\r\n'
-        '1402,Kantonale Mitteilungen / Einberufung Kantonsrat,0\r\n'
-        '1403,Kantonale Mitteilungen / Wahlen/Abstimmungen,0\r\n'
-        '1406,Kantonale Mitteilungen / Kant. Gesetzgebung,0\r\n'
-        '1411,Kantonale Mitteilungen / Mitteilungen Landschreiber,0\r\n'
-        '1412,Kantonale Mitteilungen / Kant. Stellenangebote,0\r\n'
-        '1413,Kantonale Mitteilungen / Direktion des Innern,0\r\n'
-        '1414,Kantonale Mitteilungen / Direktion für Bildung und Kultur,0\r\n'
-        '1415,Kantonale Mitteilungen / Volkswirtschaftsdirektion,0\r\n'
-        '1416,Kantonale Mitteilungen / Baudirektion,0\r\n'
-        '1418,Kantonale Mitteilungen / Gesundheitsdirektion,0\r\n'
-        '1421,Kantonale Mitteilungen / Finanzdirektion,0\r\n'
-        '1426,Kantonale Mitteilungen / Gerichtliche Bekanntmachungen,0\r\n'
-        '1427,Kantonale Mitteilungen / Konkursamt,0\r\n'
-        '16,Bürgergemeinden,3\r\n'
-        '17,Kath. Kirchgemeinden,0\r\n'
-        '18,Ev.-ref. Kirchgemeinde,0\r\n'
-        '19,Korporationen,4\r\n'
-        '20,Handelsregister,0\r\n'
     )